src/requirements.txt
click~=8.0.3
pyserial~=3.5